important changes
Important changes to index and models getter methods. diet type images not displaying fixed, ratings now display (need to make sure that the ratings are mapped from 0-10 in steps of 1) or redesign how the rating images are displayed. The recipe pictures are still static. need to migrate these to media or we cannot dynamically access recipes.